Table fixes

+++ database/migrations/2020_04_03_050146_create_global_payment_merchantid_table.php
@@ -16,7 +16,7 @@ class CreateGlobalPaymentMerchantidTable extends Migration
         Schema::create('global_payment_merchantid', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('card_type');
-            $table->integer('merchant_id');
+            $table->bigInteger('merchant_id');
             $table->timestamps();
         });
     }
